atualização

+++ CrudSapUi5/wwwroot/controller/Detalhes.controller.js
@@ -47,16 +47,25 @@ sap.ui.define([
 			}
 		},
 
-		botaoDeletar: async function (){
+		botaoDeletar:  function (){
 			let livroSelecionado = this.getView().getModel("livro").getData();
 			let idASerDeletado = livroSelecionado.codigo;
 			let oRouter = this.getOwnerComponent().getRouter();
+			MessageBox.confirm("Deseja realmente deletar esse livro?",{
+				title: "Confirmação",
+				actions: [sap.m.MessageBox.Action.OK,
+					sap.m.MessageBox.Action.CANCEL ],
 
-			await fetch(`https://localhost:7278/CrudLivro/${idASerDeletado}`, {
+				onClose: async function (oAction){
+					if(oAction === OK){
+						await fetch(`https://localhost:7278/CrudLivro/${idASerDeletado}`, {
 				method: 'DELETE'
 			})
+				oRouter.navTo("overview");
 
-			oRouter.navTo("overview");
+					}
+				}
+			})
 		},
